+++ services/preview/appplatform/mgmt/2019-05-01-preview/appplatform/CHANGELOG.md
@@ -1,11 +1,27 @@
-# Change History
+# Unreleased
 
 ## Additive Changes
 
+### New Constants
+
+1. AppResourceProvisioningState.Deleting
+1. DeploymentResourceProvisioningState.DeploymentResourceProvisioningStateDeleting
+1. TrafficDirection.Inbound
+1. TrafficDirection.Outbound
+
 ### New Funcs
 
-1. AvailableRuntimeVersions.MarshalJSON() ([]byte, error)
-1. DeploymentInstance.MarshalJSON() ([]byte, error)
-1. NetworkProfileOutboundIPs.MarshalJSON() ([]byte, error)
-1. ProxyResource.MarshalJSON() ([]byte, error)
-1. Resource.MarshalJSON() ([]byte, error)
+1. PossibleTrafficDirectionValues() []TrafficDirection
+1. RequiredTraffic.MarshalJSON() ([]byte, error)
+
+### Struct Changes
+
+#### New Structs
+
+1. RequiredTraffic
+
+#### New Struct Fields
+
+1. DeploymentResource.Sku
+1. MetricDimension.ToBeExportedForShoebox
+1. NetworkProfile.RequiredTraffics